+++ tests/Unit/Classes/ServerPermission.Tests.ps1
@@ -26,6 +26,8 @@ BeforeDiscovery {
 BeforeAll {
     $script:dscModuleName = 'SqlServerDsc'
 
+    $env:SqlServerDscCI = $true
+
     Import-Module -Name $script:dscModuleName
 
     $PSDefaultParameterValues['InModuleScope:ModuleName'] = $script:dscModuleName
@@ -40,6 +42,8 @@ AfterAll {
 
     # Unload the module being tested so that it doesn't impact any other tests.
     Get-Module -Name $script:dscModuleName -All | Remove-Module -Force
+
+    Remove-Item -Path 'env:SqlServerDscCI'
 }
 
 Describe 'ServerPermission' -Tag 'ServerPermission' {