+++ tests/Unit/Public/Uninstall-SqlDscServer.Tests.ps1
@@ -27,6 +27,8 @@ BeforeDiscovery {
 BeforeAll {
     $script:dscModuleName = 'SqlServerDsc'
 
+    $env:SqlServerDscCI = $true
+
     Import-Module -Name $script:dscModuleName
 
     $PSDefaultParameterValues['InModuleScope:ModuleName'] = $script:dscModuleName
@@ -41,6 +43,8 @@ AfterAll {
 
     # Unload the module being tested so that it doesn't impact any other tests.
     Get-Module -Name $script:dscModuleName -All | Remove-Module -Force
+
+    Remove-Item -Path 'env:SqlServerDscCI'
 }
 
 Describe 'Uninstall-SqlDscServer' -Tag 'Public' {